add wrapper around dask to extend internal Executor class; add --dask-scheduler to execute command
