+++ mapchete/cli/default/execute.py
@@ -17,6 +17,7 @@ from mapchete.cli import utils
 @utils.opt_point_crs
 @utils.opt_tile
 @utils.opt_overwrite
+@utils.opt_dask_scheduler
 @utils.opt_multi
 @utils.opt_input_file
 @utils.opt_logfile
